Eliminate the bias on the K dimension
This PR introduces eight data points, logarithmically spaced between 128 and 4096.

A few notes on evaluation: For the same problem size, the peak GFlops for Float32 is naturally twice that of Float64. An algorithm optimized for Float32 will show amplified performance when computing the average GFlops. Therefore, performance should be measured by how quickly it approaches the peak GFlops as the problem size increases.